precode.go
package main

import (
	"net/http"
	"strconv"
	"strings"
)

var cafeList = map[string][]string{
	"moscow": []string{"Мир кофе", "Сладкоежка", "Кофе и завтраки", "Сытый студент"},
}

func mainHandle(w http.ResponseWriter, req *http.Request) {
	countStr := req.URL.Query().Get("count")
	if countStr == "" {
		w.WriteHeader(http.StatusBadRequest)
		w.Write([]byte("count missing"))
		return
	}

	count, err := strconv.Atoi(countStr)
	if err != nil {
		w.WriteHeader(http.StatusBadRequest)
		w.Write([]byte("wrong count value"))
		return
	}

	city := req.URL.Query().Get("city")

	cafe, ok := cafeList[city]
	if !ok {
		w.WriteHeader(http.StatusBadRequest)
		w.Write([]byte("wrong city value"))
		return
	}

	if count > len(cafe) {
		count = len(cafe)
	}

	answer := strings.Join(cafe[:count], ",")

	w.WriteHeader(http.StatusOK)
	w.Write([]byte(answer))
}

//func TestMainHandlerWhenCountMoreThanTotal(t *testing.T) {
//	totalCount := 4
//	req := httptest.NewRequest(http.MethodGet, "/cafe?count=2&city=moscow", nil) // здесь нужно создать запрос к сервису
//	responseRecorder := httptest.NewRecorder()
//	handler := http.HandlerFunc(mainHandle)
//	handler.ServeHTTP(responseRecorder, req)
//
//	if responseRecorder.Code != http.StatusOK {
//		t.Errorf("expected status ok, got: %d", responseRecorder.Code)
//	}
//
//	// здесь нужно добавить необходимые проверки
//	assert.Equal(t, req, totalCount)
//	assert.NotEqual(t, req, totalCount)
//	assert.Len(t, req, totalCount)
//}
